requirements.txt
streamlit>=1.20
chromadb>=0.3.31
pytesseract>=0.3.10
pdfplumber>=0.7.6
python-docx>=0.8.11
Pillow>=9.0.0
pdf2image>=1.16.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.2
tqdm>=4.64.0
google-genai>=0.1.0
